+++ core/src/test/java/be/orbinson/aem/dictionarytranslator/servlets/datasource/CombiningMessageEntryDatasourceForDialogTest.java
@@ -1,11 +1,9 @@
 package be.orbinson.aem.dictionarytranslator.servlets.datasource;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.mockito.Mockito.mock;
 
 import java.io.IOException;
-import java.util.Iterator;
 import java.util.Map;
 
 import javax.servlet.RequestDispatcher;
@@ -14,7 +12,6 @@ import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 
 import org.apache.commons.collections4.IteratorUtils;
-import org.apache.jackrabbit.vault.util.Text;
 import org.apache.sling.api.request.RequestDispatcherOptions;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.SyntheticResource;
@@ -25,23 +22,23 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.osgi.util.converter.Converter;
+import org.osgi.util.converter.Converters;
 
 import com.adobe.granite.ui.components.ExpressionResolver;
 import com.adobe.granite.ui.components.ds.DataSource;
 import com.adobe.granite.ui.components.ds.EmptyDataSource;
-import com.adobe.granite.ui.components.ds.SimpleDataSource;
 import com.day.cq.replication.Replicator;
 
 import be.orbinson.aem.dictionarytranslator.services.impl.CombiningMessageEntryResourceProvider;
 import be.orbinson.aem.dictionarytranslator.services.impl.DictionaryServiceImpl;
 import io.wcm.testing.mock.aem.junit5.AemContext;
 import io.wcm.testing.mock.aem.junit5.AemContextExtension;
-
 @ExtendWith({AemContextExtension.class, MockitoExtension.class})
-class CombiningMessageEntryDatasourceTest {
+class CombiningMessageEntryDatasourceForDialogTest {
 
     private final AemContext context = new AemContext(ResourceResolverType.RESOURCEPROVIDER_MOCK);
-    CombiningMessageEntryDatasource servlet;
+    CombiningMessageEntryDatasourceForDialog servlet;
 
     @Mock
     ExpressionResolver expressionResolver;
@@ -52,14 +49,17 @@ class CombiningMessageEntryDatasourceTest {
     void beforeEach() {
         context.registerService(Replicator.class, mock(Replicator.class));
         expressionResolver = context.registerService(ExpressionResolver.class, expressionResolver);
-        context.registerInjectActivateService(new DictionaryServiceImpl());
-        context.registerInjectActivateService(new CombiningMessageEntryResourceProvider());
+        DictionaryServiceImpl dictionaryService = new DictionaryServiceImpl();
+        context.registerInjectActivateService(dictionaryService);
+        Converter converter = Converters.standardConverter();
+        CombiningMessageEntryResourceProvider.Config config = converter.convert(Map.of("enableValidation", true)).to(CombiningMessageEntryResourceProvider.Config.class);
+        context.registerInjectActivateService(new CombiningMessageEntryResourceProvider(dictionaryService, config));
         context.load().json("/content.json", "/content");
         // add additional message entry with key havgin special characters
         context.build().resource("/content/dictionaries/fruit/i18n/en/specialkey", "jcr:primaryType", "sling:MessageEntry",
                 "sling:message", "Cherry",
                 "sling:key", KEY_SPECIAL_CHARACTERS);
-        servlet = context.registerInjectActivateService(new CombiningMessageEntryDatasource());
+        servlet = context.registerInjectActivateService(new CombiningMessageEntryDatasourceForDialog());
         context.request().setRequestDispatcherFactory(new MockRequestDispatcherFactory() {
 
             @Override
@@ -90,33 +90,6 @@ class CombiningMessageEntryDatasourceTest {
     }
 
     @Test
-    void testListDataSource() throws ServletException, IOException {
-        context.request().setResource(new SyntheticResource(context.resourceResolver(), "/some/path", "artificial test resource"));
-        servlet.doGet(context.request(), context.response());
-
-        SimpleDataSource dataSource = (SimpleDataSource) context.request().getAttribute(DataSource.class.getName());
-        assertDataSourceEquals(dataSource,
-                new SyntheticResource(context.resourceResolver(), "/mnt/dictionary/content/dictionaries/fruit/i18n/apple", CombiningMessageEntryResourceProvider.RESOURCE_TYPE),
-                new SyntheticResource(context.resourceResolver(), "/mnt/dictionary/content/dictionaries/fruit/i18n/banana", CombiningMessageEntryResourceProvider.RESOURCE_TYPE),
-                new SyntheticResource(context.resourceResolver(), "/mnt/dictionary/content/dictionaries/fruit/i18n/cherry", CombiningMessageEntryResourceProvider.RESOURCE_TYPE),
-                new SyntheticResource(context.resourceResolver(), "/mnt/dictionary/content/dictionaries/fruit/i18n/" + Text.escapeIllegalJcrChars(KEY_SPECIAL_CHARACTERS), CombiningMessageEntryResourceProvider.RESOURCE_TYPE)
-        );
-    }
-
-    @Test
-    void testColumnDataSource() throws ServletException, IOException {
-        context.request().setResource(new SyntheticResource(context.resourceResolver(), "/some/path/columnsdatasource", "artificial test resource"));
-        servlet.doGet(context.request(), context.response());
-
-        DataSource dataSource = (DataSource)context.request().getAttribute(DataSource.class.getName());
-
-        assertDataSourceEquals(dataSource, CombiningMessageEntryDatasource.getColumn(context.resourceResolver(), "select", true),
-                CombiningMessageEntryDatasource.getColumn(context.resourceResolver(), "jcr:title", "Key"),
-                CombiningMessageEntryDatasource.getColumn(context.resourceResolver(), "jcr:title", "en"),
-                CombiningMessageEntryDatasource.getColumn(context.resourceResolver(), "jcr:title", "nl_BE"));
-    }
-
-    @Test
     void testItemDataSource() throws ServletException, IOException {
         context.request().setParameterMap(Map.of("item", "/mnt/dictionary/content/dictionaries/fruit/i18n/apple"));
         context.request().setResource(new SyntheticResource(context.resourceResolver(), "/some/path", "artificial test resource"));
@@ -126,27 +99,6 @@ class CombiningMessageEntryDatasourceTest {
 
         // this should only have the 3 textfields for the dialog and a hidden field for the path
         assertEquals(4, IteratorUtils.size(dataSource.iterator()));
-    }
-
-    /**
-     * As {@link DataSource} is not an iterable we cannot use {@link #assertIterableEquals(Iterable, Iterable)}.
-     * Also most resources don't have a proper implementation of {@link Object#equals()}, therefore this
-     * method compares path, resource type and (optionally) value map.
-     * @param dataSource
-     * @param expectedResources
-     */
-    void assertDataSourceEquals(DataSource dataSource, Resource... expectedResources) {
-        Iterator<Resource> iterator = dataSource.iterator();
-        int index = 0;
-        for (Resource expectedResource : expectedResources) {
-            Resource resource = iterator.next();
-            assertEquals(expectedResource.getPath(), resource.getPath(), "Path at index " + index + " does not match");
-            assertEquals(expectedResource.getResourceType(), resource.getResourceType(), "Type at index " + index + " does not match");
-            if (!expectedResource.getValueMap().isEmpty()) {
-                assertEquals(expectedResource.getValueMap(), resource.getValueMap(), "ValueMap at index " + index + " does not match");
-            }
-        }
-        assertFalse(iterator.hasNext(), () -> "Iterator has at least one more element than expected: " + iterator.next());
     }
 
 }